Add support for creating a dummy interface
Creating a dummy interface in the container's network namespace
such that a runtime could see this as a hint to look
for a vhost-user socket.

In the case of a Clear Container, this will trigger
cc-runtime to search the filesystem for the socket and
pass it to the continer hypervisor accordingly.

Signed-off-by: Eric Ernst <eric.ernst@intel.com>